# data_structures: make TinyList more readable and optimize remove(_)